code/Source-credibility/requirements.txt
requests
flask
flask_cors
numpy
coloredlogs
tldextract==3.1.0
sourceRank
kafka-python
elasticsearch
elasticsearch_dsl
python-Levenshtein
fuzzywuzzy